% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moving_average_plot.R
\name{add_moving_average_plot}
\alias{add_moving_average_plot}
\title{Add a Moving Average Line to a Plotly Time Series Plot}
\usage{
add_moving_average_plot(
  p,
  moving_avg,
  df,
  time_var,
  value_var,
  k_ma,
  color = dict_color[["moving_avg_color"]]
)
}
\arguments{
\item{p}{A plotly object to which the moving average line will be added.}

\item{moving_avg}{Logical; whether to add the moving average line.}

\item{df}{A data frame containing the input time series data.}

\item{time_var}{Character; name of the time variable column.}

\item{value_var}{Character; name of the numeric value column to smooth.}

\item{k_ma}{Integer; number of periods for the moving average.}

\item{color}{Character; color for the moving average line.}
}
\value{
A plotly object with the moving average line added (if enabled).
}
\description{
Computes and overlays a moving average line on an existing Plotly time series plot.
This helps visualize longer-term trends by smoothing short-term fluctuations.
}
\examples{
df <- data.frame(time = Sys.Date() + 1:10, value = rnorm(10))
p <- plotly::plot_ly(df, x = ~time, y = ~value, type = "scatter", mode = "lines")
add_moving_average_plot(p, TRUE, df, "time", "value", 3)
}
