% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex_functions.R
\name{apply_plot_pipeline}
\alias{apply_plot_pipeline}
\title{Apply a Pipeline of Plotly Transformations}
\usage{
apply_plot_pipeline(p, pipeline)
}
\arguments{
\item{p}{A plotly object (e.g., returned from \code{plot_ly()} or \code{ggplotly()}).}

\item{pipeline}{A list of functions, each taking and returning a plotly object.}
}
\value{
A plotly object resulting from applying all transformations in sequence.
}
\description{
This function applies a sequence of transformation functions to a plotly plot object.
Each function in the pipeline must take a plotly object as input and return a modified one.
This enables composable, modular manipulation of plots (e.g., adding trend lines, overlays).
}
\examples{
\dontrun{
pipeline <- list(
  function(p) add_trend_line(p, trend = TRUE, data = df, time_var = time, value_var = value),
  function(p) add_moving_average_plot(p, TRUE, df, value_var = value, k_ma = 3)
)
base_plot <- plot_ly(df, x = ~time) \%>\% add_bars(y = ~value)
result_plot <- apply_plot_pipeline(base_plot, pipeline)
}
}
\keyword{internal}
