% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex_functions.R
\name{build_all_plot_inputs}
\alias{build_all_plot_inputs}
\title{Build list of plot input parameters for all schema-defined variables}
\usage{
build_all_plot_inputs(df, schema)
}
\arguments{
\item{df}{A data frame containing the time series data to be plotted.}

\item{schema}{A schema list created by \code{generate_plot_schema()}, containing global and per-variable configs.}
}
\value{
A named list of argument lists (one per plot variable) for use with \code{do.call()}.
}
\description{
Iterates over all variables defined in the \code{schema$plots} section and constructs
individual argument lists for each using \code{build_plot_args()}. This is a helper function
that prepares inputs for batch plotting functions like \code{get_combined_plots()}.
}
\examples{
\dontrun{
schema <- generate_plot_schema(input)
df <- prepare_data_from_schema(schema)
input_list <- build_all_plot_inputs(df, schema)
}
}
