% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{dict_color}
\alias{dict_color}
\title{Global color dictionary for plotting}
\format{
An object of class \code{list} of length 8.
}
\usage{
dict_color
}
\description{
This configuration list defines consistent colors used across plotting functions.
}
\keyword{datasets}
