% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_window_default.R
\name{set_last_window_view}
\alias{set_last_window_view}
\title{Add a Custom "Restore 10-Year View" Button to Plotly Modebar}
\usage{
set_last_window_view(p, df, time_var)
}
\arguments{
\item{p}{A Plotly object (e.g., output from plot_ly()).}

\item{df}{A data.frame containing the data used in the plot.}

\item{time_var}{A string representing the column name in \code{df} with time data (must be POSIXct or Date).}
}
\value{
A Plotly object with an embedded JavaScript modebar button to reset the x-axis range.
}
\description{
This function adds a custom button to the Plotly modebar that resets the x-axis
to display the last 10 years of data. It dynamically detects the maximum date
in the given data frame and sets the view range accordingly.
}
\examples{
library(plotly)
library(lubridate)
df <- data.frame(date = seq(ymd("2000-01-01"), ymd("2040-01-01"), by = "months"),
                 value = rnorm(481))
p <- plot_ly(df, x = ~date, y = ~value, type = "scatter", mode = "lines")
p <- set_last_window_view(p, df, "date")
p

}
